Add credentials.conf in CMakeLists
This commit automatically adds the credentials.conf file from the root
app directory at build time. If that file is missing, the build will
fail and issue an error message about the missing file. This is ideal as
it prevents user from running apps without adding WiFi and Golioth
credentials.

Signed-off-by: Mike Szczys <mike@golioth.io>